requirements.txt
streamlit 
google-generativeai 
requests 
korean_lunar_calendar 
python-dotenv 
pinecone-client